% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggite2.R
\name{aggite2}
\alias{aggite2}
\title{Pair Aggregate Unit-Time Average Treatment Effects}
\usage{
aggite2(
  MP,
  type = "group",
  type2 = "dynamic",
  balance_e = NULL,
  min_e = -Inf,
  max_e = Inf,
  na.rm = FALSE,
  indep = TRUE,
  bstrap = NULL,
  biters = NULL,
  cband = NULL,
  alp = NULL,
  clustervars = NULL
)
}
\arguments{
\item{MP}{an MP_i object (i.e., the results of the \code{\link[=att_it]{att_it()}} method)}

\item{type}{Only allows "group" (default) and custom aggregators.}

\item{type2}{Only allows "dynamic".}

\item{balance_e}{If set (and if one computes dynamic effects), it balances
the sample with respect to event time.  For example, if \code{balance.e=2},
\code{aggite} will drop groups that are not exposed to treatment for
at least three periods. (the initial period when \code{e=0} as well as the
next two periods when \code{e=1} and the \code{e=2}).  This ensures that
the composition of groups does not change when event time changes.}

\item{min_e}{For event studies, this is the smallest event time to compute
dynamic effects for.  By default, \code{min_e = -Inf} so that effects at
all lengths of exposure are computed.}

\item{max_e}{For event studies, this is the largest event time to compute
dynamic effects for.  By default, \code{max_e = Inf} so that effects at
all lengths of exposure are computed.}

\item{na.rm}{Logical value. If TRUE, removes missing values from analyses. Default is FALSE.}

\item{indep}{Logical value. If TRUE, cross-sectional units are aggregated under an independence assumption. Results in a tighter bound. Default is TRUE.}

\item{bstrap}{Does not apply. To control aggregation assumptions see \code{indep}.}

\item{biters}{Does not apply.}

\item{cband}{Does not apply. Set \code{indep = FALSE} for a conservative bound.}

\item{alp}{the significance level, default is value set in the MP_i object.}

\item{clustervars}{Does not apply.}
}
\value{
An \code{\link{AGGITEobj}} object that holds the results from the aggregation.
}
\description{
A function to take unit-time average treatment effects
and aggregate them into sub-effects where there is more than one aggregating type.
Possible aggregations include pairs of any of the following:
"group" or a custom aggregation name with "dynamic".
}
\examples{
# first run the att_it() function
simdata = sim_data(posttreat=6)
attobject = att_it(yname = "y", tname = "time", gname = "treatg", idname ="unit", data = simdata)

# aggregate all post-treatment effects of the group and dynamic level
agtobject = aggite2(attobject,type="group",type2="dynamic")



}
