% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGGITEobj.R
\name{AGGITEobj}
\alias{AGGITEobj}
\title{Aggregate Treatment Effect Parameters Object}
\usage{
AGGITEobj(
  overall.att = NULL,
  overall.se = NULL,
  overall.lci = NULL,
  overall.uci = NULL,
  overall.count = NULL,
  type = "simple",
  type2 = NULL,
  egt = NULL,
  egt2 = NULL,
  att.egt = NULL,
  se.egt = NULL,
  lci.egt = NULL,
  uci.egt = NULL,
  count.egt = NULL,
  crit.val.egt = NULL,
  inf.function = NULL,
  min_e = NULL,
  max_e = NULL,
  balance_e = NULL,
  indep = NULL,
  call = NULL,
  DIDparams = NULL
)
}
\arguments{
\item{overall.att}{The estimated overall ATT}

\item{overall.se}{Standard error for overall ATT. Not reported for conformal intervals.}

\item{overall.lci}{The lower confidence interval of the overall ATT}

\item{overall.uci}{The upper confidence interval of the overall ATT}

\item{overall.count}{The number of cross sectional units used to produce the overall ATT}

\item{type}{Only allows "group" (default) and custom aggregators.}

\item{type2}{Only allows "dynamic".}

\item{egt}{Holds the length of exposure (for dynamic effects), the group, the unit,
cohort, custom aggregator, or the time period (for calendar time effects)}

\item{egt2}{a second aggregation type to hold a second-level aggregator, NULL if no such object}

\item{att.egt}{The ATT specific to egt}

\item{se.egt}{The standard error specific to egt from the bootstrap. Not reported for conformal intervals.}

\item{lci.egt}{The lower confidence interval}

\item{uci.egt}{The upper confidence interval}

\item{count.egt}{The number of cross-sectional units used to produce the egt.}

\item{crit.val.egt}{Always NULL because simultaneous option is turned off.}

\item{inf.function}{The bootstrap draws specific to the egt. Name is carried over from the did package.}

\item{min_e}{For event studies, this is the smallest event time to compute
dynamic effects for.  By default, \code{min_e = -Inf} so that effects at
all lengths of exposure are computed.}

\item{max_e}{For event studies, this is the largest event time to compute
dynamic effects for.  By default, \code{max_e = Inf} so that effects at
all lengths of exposure are computed.}

\item{balance_e}{If set (and if one computes dynamic effects), it balances
the sample with respect to event time.  For example, if \code{balance.e=2},
\code{aggite} will drop groups that are not exposed to treatment for
at least three periods. (the initial period when \code{e=0} as well as the
next two periods when \code{e=1} and the \code{e=2}).  This ensures that
the composition of groups does not change when event time changes.}

\item{indep}{Logical value. If TRUE, cross-sectional units are aggregated under an independence assumption. Results in a tighter bound. Default is TRUE.}

\item{call}{The function call to aggite2}

\item{DIDparams}{A DIDparams_i object}
}
\value{
an AGGITEobj
}
\description{
Aggregate Treatment Effect Parameters Object
}
\examples{
# A helper function for [aggite()]. See examples in documentation in that function.


}
