% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drreg_att.R
\name{drreg_att}
\alias{drreg_att}
\title{The average treatment effect calculated using doubly-robust regression.}
\usage{
drreg_att(
  index,
  y1,
  y0,
  D,
  covariates,
  i.weights = NULL,
  dist.family = stats::gaussian(link = "identity"),
  alpha = 0.05,
  conformal_split = 1,
  weighted_conformal = FALSE
)
}
\arguments{
\item{index}{a numeric vector that identifies each unit.}

\item{y1}{the outcome vector of the post-treatment period.}

\item{y0}{the outcome vector of the pre-treatment period.}

\item{D}{a binary vector indicating treatment.}

\item{covariates}{a matrix of covariates, which must include covariates.}

\item{i.weights}{sample weights, not to be confused with inverse propensity weights.}

\item{dist.family}{the distribution of the outcome variable. Defaults to gaussian.}

\item{alpha}{the confidence level for the interval estimates. Defaults to 0.05.}

\item{conformal_split}{the size of the holdout set when cross-fitting. Defaults to jackknife+ (i.e. 1)}

\item{weighted_conformal}{whether the conformal scores should be weighted}
}
\value{
a list containing the calculated estimate \code{ATT}, the conformal interval \code{lci} and \code{uci} and a matrix containing the conformal scores and inverse propensity weights.
}
\description{
The average treatment effect calculated using doubly-robust regression.
}
\examples{
# A dataset with a single treated unit
data <- sim_data(10,10,1,1, cohort = rep(1,21), pretreat = 1, posttreat = 0)
datapre <- data[data$time==9,]
data <- data[data$time==10,]
# Run the function
drobject <- drreg_att(index=data$unit,y1=data$y,y0=datapre$y,D=data$posttreat, covariates = NULL)
# Report the ATT
drobject$ATT
}
