% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_i.R
\name{MP_i}
\alias{MP_i}
\title{MP_i}
\usage{
MP_i(
  id,
  group,
  t,
  att,
  lci,
  uci,
  CS,
  IPW,
  M,
  n = NULL,
  aggite = NULL,
  alp = 0.05,
  ipwqual = NULL,
  attcalc = NULL,
  baseline = NULL,
  deltaY = NULL,
  count = NULL,
  DIDparams = NULL
)
}
\arguments{
\item{id}{which unit the unit-time treatment effects are for}

\item{group}{which group (defined by period first treated) an unit-time average treatment effect is for}

\item{t}{which time period a group-time average treatment effect is for}

\item{att}{the estimate for unit \code{id} and time
period \code{t}}

\item{lci}{lower confidence interval for att}

\item{uci}{upper confidence interval for att}

\item{CS}{the matrix of conformal scores to produce the aggregate intervals}

\item{IPW}{the matrix of weights for weighted conformal intervals}

\item{M}{the matrix of counterfactual predictions}

\item{n}{the number of unique cross-sectional units (unique values of idname)}

\item{aggite}{an aggregate treatment effects object}

\item{alp}{the significance level, default is 0.05}

\item{ipwqual}{the maximum propensity score}

\item{attcalc}{similar to att but does not remove estimates when there are propensity score problems}

\item{baseline}{the baseline of the id}

\item{deltaY}{the outcome difference from the baseline of the id}

\item{count}{takes a count of the number of units in the component estimate}

\item{DIDparams}{a \code{\link{DIDparams_i}} object.
of the call to \code{\link[=att_it]{att_it()}}.}
}
\value{
MP_i object
}
\description{
Multi-period objects that hold results for unit-time average treatment effects
}
\examples{
# Helper function for [att_it()]. See documentation of that function for an example.


}
