% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_attit.R
\name{process_attit}
\alias{process_attit}
\title{Process Results from \code{\link[=compute.att_it]{compute.att_it()}}}
\usage{
process_attit(attgt.list)
}
\arguments{
\item{attgt.list}{list of results from \code{\link[=compute.att_it]{compute.att_it()}}}
}
\value{
list with elements:
\item{id}{which id a set of results belongs to}
\item{group}{which group the id belongs to}
\item{tt}{which time period a set of results belongs to}
\item{att}{the unit time average treatment effect}
\item{lci}{the lower confidence interval of the unit time average treatment effect}
\item{uci}{the upper confidence interval of the unit time average treatment effect}
\item{ipwqual}{the maximum propensity score if est_method is ipw or dr}
\item{baseline}{the pre-treatment outcome value of corresponding id}
\item{deltaY}{the difference in outcome from pre-treatment of corresponding id}
\item{attcalc}{the unit time average treatment effect}
}
\description{
Process Results from \code{\link[=compute.att_it]{compute.att_it()}}
}
\examples{
# A helper function of [att_it()], but can be useful for debugging.


}
