% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_quantile.R
\name{weighted_quantile}
\alias{weighted_quantile}
\title{A function to get weighted quantiles}
\usage{
weighted_quantile(x, w = NULL, alpha, lower = TRUE)
}
\arguments{
\item{x}{a numeric vector of draws}

\item{w}{a numeric vector of weights the same length as draws}

\item{alpha}{the miscoverage level.}

\item{lower}{if TRUE, the lower quantile is reported. Default is TRUE.}
}
\value{
a numeric estimate of the lower or upper quantile
}
\description{
A function to get weighted quantiles
}
\examples{
weighted_quantile(rnorm(100,0,1),alpha=0.05)


}
