% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_quantile.R
\name{weighted_sd}
\alias{weighted_sd}
\title{A function to get weighted standard deviation}
\usage{
weighted_sd(x, w = NULL, na.rm = FALSE, unbiased = TRUE)
}
\arguments{
\item{x}{a numeric vector of draws}

\item{w}{a numeric vector of weights the same length as draws}

\item{na.rm}{if TRUE, ignores missing values. Default is TRUE.}

\item{unbiased}{if TRUE, a Bessel-type correction is applied. Default is TRUE.}
}
\value{
a numeric estimate of standard deviation
}
\description{
A function to get weighted standard deviation
}
\examples{
weighted_sd(rnorm(100,0,1))
}
